Add pipes.c
This program uses two pipes to demonstrate bidirectional communication between a parent process and a child process. The parent process accepts characters from stdin 1 byte at a type and pipes the data to the child process. The child process then converts the characters to upper case and pipes the data back to the parent process, which then prints the uppercase text to stdout.