Add fm359_signal.c
This program simulates the linux kill command. It takes process IDs as user input and sends a signal to them as specified by the user. The user has the option to select which signal is sent as an input argument along with the -s flag. The -l displays all the available signals. And the program sends the SIGTERM signal by default, when no flag is specified.