Add fm359_primes.c
This program computes the first 10,000,000 prime numbers and stores them in an array. Using SIGALRM the program prints the last five prime numbers found and displays them to the terminal as well as the number of prime numbers computed at the time. It also accepts SIGUSR1 to display the last 5 primes upon receiving the signal at any time from the user. Finally, it accepts SIGTERM and displays the current status before exiting the program.