Add fm359_hw5.c
This program computes the sum of the squares of the first 20 integers from 0 to 19 using multithreading and compares the result with a single threaded loop to calculate the same summation. The program makes use of a control variable to ensure that the  main loop does not terminate before all of the threads have finished executing.